make renderStrategy required
